Move type set to its own file and document.
